+++ aiko/src/services/equipment/index.ts
@@ -1,0 +1,11 @@
+import equipment from "../../../data/equipment.json"
+
+export default function equipmentServices() {
+    function getEquipments() {
+        return equipment
+    }
+
+    return {
+        getEquipments,
+    }
+}
