+++ aiko/src/services/equipmentModelServices/index.ts
@@ -1,0 +1,11 @@
+import equipmentModel from "../../../data/equipmentModel.json"
+
+export default function equipmentModelServices() {
+    function getEquipmentsModel() {
+        return equipmentModel
+    }
+
+    return {
+        getEquipmentsModel,
+    }
+}
