+++ aiko/src/services/equipmentsPositionHistoryServices/index.ts
@@ -1,0 +1,11 @@
+import equipmentPositionHistory from "../../../data/equipmentPositionHistory.json"
+
+export default function equipmentsPositionHistoryServices() {
+    function getEquipmentsPositionHistory() {
+        return equipmentPositionHistory
+    }
+
+    return {
+        getEquipmentsPositionHistory,
+    }
+}
