+++ aiko/src/services/equipmentsStateHistoryServices/index.ts
@@ -1,0 +1,11 @@
+import equipmentStateHistory from "../../../data/equipmentStateHistory.json"
+
+export default function equipmentsStateHistoryServices() {
+    function getEquipmentsStateHistory() {
+        return equipmentStateHistory
+    }
+
+    return {
+        getEquipmentsStateHistory,
+    }
+}
