+++ aiko/src/services/equipmentsStateServices/index.ts
@@ -1,0 +1,11 @@
+import equipmentState from "../../../data/equipmentState.json"
+
+export default function equipmentsStateServices() {
+    function getEquipmentsState() {
+        return equipmentState
+    }
+
+    return {
+        getEquipmentsState,
+    }
+}
